Fix decline as per Daniel comments separate accept from dojo.js
