+++ dojos.js
@@ -103,10 +103,10 @@ module.exports = function (options) {
   seneca.add({role: plugin, cmd: 'request_user_invite'}, cmd_request_user_invite);
   seneca.add({role: plugin, cmd: 'load_dojo_champion'}, cmd_load_dojo_champion);
   seneca.add({role: plugin, cmd: 'search_join_requests'}, cmd_search_join_requests);
-  seneca.add({role: plugin, cmd: 'accept_user_request'}, cmd_accept_user_request);
-  seneca.add({role: plugin, cmd: 'decline_join_request'}, require('./lib/dojos/decline-join-request').bind(seneca));
+  seneca.add({role: plugin, cmd: 'accept_user_request'}, require('./lib/dojos/accept-join-request'));
+  seneca.add({role: plugin, cmd: 'decline_join_request'}, require('./lib/dojos/decline-join-request'));
   seneca.add({role: plugin, cmd: 'dojos_for_user'}, cmd_dojos_for_user);
-  seneca.add({role: plugin, cmd: 'save_usersdojos'}, require('./lib/save-usersdojo').bind(seneca));
+  seneca.add({role: plugin, cmd: 'save_usersdojos'}, require('./lib/save-usersdojo'));
   seneca.add({role: plugin, cmd: 'remove_usersdojos'}, cmd_remove_usersdojos);
   seneca.add({role: plugin, cmd: 'get_user_types'}, cmd_get_user_types);
   seneca.add({role: plugin, cmd: 'get_user_permissions'}, cmd_get_user_permissions);
@@ -124,7 +124,7 @@ module.exports = function (options) {
   seneca.add({role: plugin, cmd: 'find_dojolead'}, cmd_find_dojolead);
   seneca.add({role: plugin, cmd: 'load_dojo_email'}, cmd_load_dojo_email);
   seneca.add({role: plugin, cmd: 'notify_all_members'}, cmd_notify_all_members);
-  seneca.add({role: plugin, cmd: 'add_children_parent_dojo'}, addChildrenParentDojo.bind(seneca));
+  seneca.add({role: plugin, cmd: 'add_children_parent_dojo'}, addChildrenParentDojo);
 
   //  TODO:10 : export as a different microservice?
   //  Polls channels
@@ -1721,97 +1721,6 @@ module.exports = function (options) {
         done(null, champions);
       });
     });
-  }
-
-  function cmd_accept_user_request (args, done) {
-    var tokenData = args.data;
-    if (!args.user) return done();
-    var currentUserId = args.user.id;
-    var inviteTokenId = tokenData.inviteToken;
-    var requestedByUser = tokenData.requestedByUser;
-
-    async.waterfall([
-      loadUser,
-      verifyRequest,
-      updateUser,
-      tidyUpJoinRequests
-    ], function (err, res) {
-      if (err) return done(null, {error: err.message});
-      return done(null, res);
-    });
-
-    function loadUser (done) {
-      seneca.act({role: 'cd-users', cmd: 'load', id: requestedByUser, user: args.user}, function (err, response) {
-        if (err) return done(err);
-        var user = response;
-        var joinRequests = user.joinRequests;
-        var validRequestFound = _.find(joinRequests, {id: inviteTokenId});
-        if (!validRequestFound) return done(new Error('Join request not found'));
-        return done(null, validRequestFound);
-      });
-    }
-
-    function verifyRequest (validRequestFound, done) {
-      // Make sure current user is dojo admin.
-      var dojoId = validRequestFound.dojoId;
-      seneca.act({
-        role: plugin,
-        cmd: 'load_usersdojos',
-        query: {userId: currentUserId, dojoId: dojoId}
-      }, function (err, usersDojos) {
-        if (err) return done(err);
-        var dojoAdminError = new Error('You must have the Dojo Admin permission to accept user join requests.');
-        if (!usersDojos.length) return done(dojoAdminError);
-        var userDojo = usersDojos[0];
-        var dojoAdminPermissionFound = _.find(userDojo.userPermissions, function (userPermission) {
-          return userPermission.name === 'dojo-admin';
-        });
-        if (!dojoAdminPermissionFound) return done(dojoAdminError);
-        return done(null, validRequestFound);
-      });
-    }
-
-    function updateUser (joinRequest, done) {
-      // Add type to userTypes in cd/usersdojos.
-      // Add user to dojo users if not already added.
-      var userDojo = {};
-      seneca.act({
-        role: plugin,
-        cmd: 'load_usersdojos',
-        query: {userId: requestedByUser, dojoId: joinRequest.dojoId}
-      }, function (err, response) {
-        if (err) return done(err);
-        if (_.isEmpty(response)) {
-          userDojo.owner = 0;
-          userDojo.userId = requestedByUser;
-          userDojo.dojoId = joinRequest.dojoId;
-          userDojo.userTypes = [];
-          userDojo.userTypes.push(joinRequest.userType);
-        } else {
-          // Update cd/usersdojos
-          userDojo = response[0];
-          if (!userDojo.userTypes) userDojo.userTypes = [];
-          userDojo.userTypes.push(joinRequest.userType);
-        }
-        // If invite token user type is champion, update user permissions
-        if (joinRequest.userType === 'champion') {
-          userDojo.userPermissions = [
-            {title: 'Dojo Admin', name: 'dojo-admin'},
-            {title: 'Ticketing Admin', name: 'ticketing-admin'}
-          ];
-        }
-        seneca.act({role: plugin, cmd: 'save_usersdojos', userDojo: userDojo, user: args.user, invite: joinRequest}, done);
-      });
-    }
-
-    function tidyUpJoinRequests (userDojo, done) {
-      seneca.act({role: 'cd-users', cmd: 'load', id: requestedByUser, user: args.user}, function (err, user) {
-        if (err) return done(err);
-        user.joinRequests = _.without(user.joinRequests, _.find(user.joinRequests, {id: inviteTokenId}));
-        var toBeUpdated = _.pick(user, ['id', 'joinRequests']);
-        seneca.act({role: 'cd-users', cmd: 'update', user: toBeUpdated, id: user.id}, done);
-      });
-    }
   }
 
   function cmd_dojos_for_user (args, done) {
