+++ lib/dojos/decline-join-request.js
@@ -3,25 +3,24 @@ var async = require('async');
 function cmd_accept_user_request (args, done) {
   var seneca = this;
   var plugin = args.role;
-  var tokenData = args.data;
+  if (!args.user) return done(new Error('Missing parameter for cmd_accept_user_request'));
   var currentUserId = args.user.id;
-  if (!args.user) return done();
-  var inviteTokenId = tokenData.inviteToken;
-  var requestedByUser = tokenData.requestedByUser;
-
+  var inviteTokenId = args.inviteToken;
+  var requestedByUser = args.requestedByUser;
+  var user = {};
   async.waterfall([
-    loadUser,
-    verifyRequest,
+    verifyJoinRequest,
+    checkPerms,
     removeJoinRequests
   ], function (err, res) {
     if (err) return done(null, {error: err.message});
     return done(null, res);
   });
 
-  function loadUser (done) {
+  function verifyJoinRequest (done) {
     seneca.act({role: 'cd-users', cmd: 'load', id: requestedByUser, user: args.user}, function (err, response) {
       if (err) return done(err);
-      var user = response;
+      user = response;
       var joinRequests = user.joinRequests;
       var validRequestFound = _.find(joinRequests, {id: inviteTokenId});
       if (!validRequestFound) return done(new Error('Join request not found'));
@@ -29,7 +28,7 @@ function cmd_accept_user_request (args, done) {
     });
   }
 
-  function verifyRequest (validRequestFound, done) {
+  function checkPerms (validRequestFound, done) {
     // Make sure current user is dojo admin.
     var dojoId = validRequestFound.dojoId;
     seneca.act({
@@ -49,13 +48,10 @@ function cmd_accept_user_request (args, done) {
     });
   }
 
-  function removeJoinRequests (done) {
-    seneca.act({role: 'cd-users', cmd: 'load', id: requestedByUser, user: args.user}, function (err, user) {
-      if (err) return done(err);
-      user.joinRequests = _.without(user.joinRequests, _.find(user.joinRequests, {id: inviteTokenId}));
-      var toBeUpdated = _.pick(user, ['id', 'joinRequests']);
-      seneca.act({role: 'cd-users', cmd: 'update', user: toBeUpdated, id: user.id}, done);
-    });
+  function removeJoinRequests (validRequestFound, done) {
+    user.joinRequests = _.without(user.joinRequests, validRequestFound);
+    var toBeUpdated = _.pick(user, ['id', 'joinRequests']);
+    seneca.act({role: 'cd-users', cmd: 'update', user: toBeUpdated, id: user.id}, done);
   }
 }
