+++ lib/dojos/search-join-requests.js
@@ -8,25 +8,29 @@ var _ = require('lodash');
 module.exports = function (args, done) {
   var seneca = this;
   var dojoId = args.dojoId;
-  var query = args.query || {};
-  if (query.name) {
-    var nameQuery = query.name;
-    delete query.name;
-  }
-  query.dojoId = dojoId;
+  var query = {};
   if (!dojoId) return done(new Error('Missing parameter DojoId'));
+  query.dojoId = dojoId;
+  // Optional params
+  if (args.name) {
+    var nameQuery = args.name;
+  }
+  if (args.userType) {
+    query.userType = args.userType;
+  }
+
   seneca.act({role: 'cd-users', domain: 'join_requests', cmd: 'search', query: query},
     function (err, requests) {
       if (err) return done(err);
       async.mapSeries(requests, function (request, cb) {
-        var lQuery = _.clone(query);
-        lQuery.userId = request.userId;
+        var lQuery = {
+          userId: request.userId
+        };
         seneca.act({role: 'cd-profiles', cmd: 'user_profile_data', query: lQuery, user: args.user}, function (err, userData) {
           if (err) cb(err);
+          request.userData = userData;
           if (nameQuery) {
             if (!userData.name.match(nameQuery)) request = null;
-          } else {
-            request.userData = userData;
           }
           cb(null, request);
         });
